Refactor mob drops to own package
