+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/registry/PylonRegistryKey.kt
@@ -3,7 +3,7 @@ package io.github.pylonmc.pylon.core.registry
 import io.github.pylonmc.pylon.core.addon.PylonAddon
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema
 import io.github.pylonmc.pylon.core.item.PylonItemSchema
-import io.github.pylonmc.pylon.core.recipe.MobDrop
+import io.github.pylonmc.pylon.core.mobdrop.MobDrop
 import io.github.pylonmc.pylon.core.recipe.RecipeType
 import io.github.pylonmc.pylon.core.test.GameTestConfig
 import io.github.pylonmc.pylon.core.util.pylonKey
