+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/recipe/MobDropTest.java
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.test.test.recipe;
 
-import io.github.pylonmc.pylon.core.recipe.SimpleMobDrop;
+import io.github.pylonmc.pylon.core.mobdrop.SimpleMobDrop;
 import io.github.pylonmc.pylon.core.registry.PylonRegistry;
 import io.github.pylonmc.pylon.core.test.GameTestConfig;
 import io.github.pylonmc.pylon.test.PylonTest;
